+++ changelog.md
@@ -1,7 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+
+## [git] - 2021-01-21
+### Added
+- support `-f` option for fahrenheit in get_temp.sh and read-loop.sh.
+
+### Changed
+- Move business logic and related variables including settings to a class.
+- Move settings to a dict named `settings` and rename them as follows:
+  - `desired_degrees` to `target`.
+  - `desired_format` to `scale`.
+  - `interval_seconds` to `interval`.
+  - `desired_total_seconds` to `minTime`.
+  - `desired_comparison` to `compareOp`.
+  - `desired_collate_method` to `useStat`.
+- Use ccwienk/temper in TemperatureSanitizer.py for improved compatibility.
 
 
 ## [git] - 2021-01-20
@@ -10,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   only in get_temp.sh (which read-loop.sh uses if hid-query isn't
   present)
   - Tested with 413d:2107 (applies to [Jepeak Temper High Accurate USB Thermometer Temperature Sensor Data Logger Record for PC Laptop](https://www.amazon.com/gp/product/B009YRP906/ref=ppx_yo_dt_b_search_asin_title?ie=UTF8&psc=1) and others)
+
 ### Changed
 - Move changes to Changelog.
 
@@ -24,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Add a region called User Settings to distinguish those variables from
   the rest of the code.
+
 ### Changed
 - Renamed variables for clarity